Struct shred::AsyncDispatcher

pub struct AsyncDispatcher<'a, R> { /* fields omitted */ }

Like, Dispatcher but works asynchronously.

Methods

impl<'a, R> AsyncDispatcher<'a, R> where
R: Borrow<Resources> + Send + Sync + 'static,

pub fn dispatch(&mut self)

Dispatches the systems asynchronously. Does not execute thread local systems.

If you want to wait for the systems to finish, call wait().

pub fn wait(&mut self)

Waits for all the asynchronously dispatched systems to finish and executes thread local systems (if there are any).

pub fn wait_without_tl(&mut self)

Waits for all the asynchronously dispatched systems to finish without executing thread local systems.

pub fn is_running(&self) -> bool

Checks if any of the asynchronously dispatched systems are running.

pub fn dispatch_thread_local(&mut self)

Dispatch only thread local systems sequentially.

If wait_without_tl() or wait() wasn't called before, this method will wait.

pub fn mut_res(&mut self) -> &mut R

Returns the resources.

If wait_without_tl() or wait() wasn't called before, this method will do that.