Struct shred::DefaultBatchControllerSystem

pub struct DefaultBatchControllerSystem<'a, 'b> { /* fields omitted */ }

The DefaultBatchControllerSystem is a simple implementation that will dispatch the inner dispatcher one time.

Usually you want to create your own Dispatcher.

Is safe to implement Send and Sync because the BatchAccessor keep tracks of all used resources and thus the System can be safely executed in multi thread.