Struct shred::Seq

pub struct Seq<H, T> { /* fields omitted */ }

Runs two tasks sequentially. These two tasks are called head and tail in the following documentation.

Methods

impl<H> Seq<H, Nil>

pub fn new(head: H) -> Self

Creates a new Seq struct, with the tail being a no-op.

pub fn with<T>(self, sys: T) -> Seq<Seq<H, T>, Nil>

Adds sys as the second job and returns a new Seq struct with the previous struct as head and a no-op tail.