Type Definition shred::ReadExpect

type ReadExpect<'a, T> = Read<'a, T, PanicHandler>;

Allows to fetch a resource in a system immutably. This will panic if the resource does not exist. Usage of Read or Option<Read> is therefore recommended.