Struct shred::AsyncDispatcher

pub struct AsyncDispatcher<'a, R> { /* fields omitted */ }

Like, Dispatcher but works asynchronously.

Methods

impl<'a, R> AsyncDispatcher<'a, R> where
R: Borrow<Resources> + Send + Sync + 'static,

pub fn dispatch(&mut self)

Dispatches the systems asynchronously. Does not execute thread local systems.

If you want to wait for the systems to finish, call wait().

pub fn wait(&mut self)

Waits for all the asynchronously dispatched systems to finish and executes thread local systems (if there are any).

pub fn wait_without_tl(&mut self)

Waits for all the asynchronously dispatched systems to finish without executing thread local systems.

See wait for executing thread local systems.

pub fn running(&mut self) -> bool

Checks if any of the asynchronously dispatched systems are running.

pub fn res(&mut self) -> &R

Returns the resources.

This will wait for the asynchronous systems to finish.

pub fn mut_res(&mut self) -> &mut R

Returns the resources mutable.

This will wait for the asynchronous systems to finish.