Trait shred::Resource

pub trait Resource: Any + Send + Sync + 'static { }

A resource is a data slot which lives in the World can only be accessed according to Rust's typical borrowing model (one writer xor multiple readers).

Methods

impl dyn Resource

pub fn is<T: Resource>(&self) -> bool

Returns true if the boxed type is the same as T

pub fn downcast_ref<T: Resource>(&self) -> Option<&T>

Returns some reference to the boxed value if it is of type T, or None if it isn't.

pub unsafe fn downcast_ref_unchecked<T: Resource>(&self) -> &T

Returns a reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

pub fn downcast_mut<T: Resource>(&mut self) -> Option<&mut T>

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

pub unsafe fn downcast_mut_unchecked<T: Resource>(&mut self) -> &mut T

Returns a mutable reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

impl dyn Resource

pub fn downcast<T: Resource>(self: Box<Self>) -> Result<Box<T>, Box<Self>>

Returns the boxed value if it is of type T, or Err(Self) if it isn't.

pub unsafe fn downcast_unchecked<T: Resource>(self: Box<Self>) -> Box<T>

Returns the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.